Add 2024 problem 13 solution part 2

+++ Advent2024/Problem13/Vector.cs
@@ -1,3 +1,3 @@
 namespace Advent2024.Problem13;
 
-public record Vector(int X, int Y);
+public record Vector(long X, long Y);
